refactor: introduce commit sync service
